@@ -50,7 +50,8 @@
     "path-exists": "2.1.0",
     "postcss-loader": "^1.0.0",
     "promise": "7.1.1",
-    "react": "^15.4.1",
+    "prop-types": "^15.5.10",
+    "react": "^15.6.1",
     "react-datetime": "^2.8.1",
     "react-dev-utils": "^0.4.0",
     "react-dom": "^15.4.1",
