@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 
 export default function Elevation({ children, zLevel }) {
   return (
@@ -9,17 +10,17 @@ export default function Elevation({ children, zLevel }) {
 }
 
 Elevation.propTypes = {
-  children: React.PropTypes.oneOfType([
-    React.PropTypes.element,
-    React.PropTypes.node,
-    React.PropTypes.arrayOf(
-      React.PropTypes.oneOfType([
-        React.PropTypes.element,
-        React.PropTypes.arrayOf(React.PropTypes.element)
+  children: PropTypes.oneOfType([
+    PropTypes.element,
+    PropTypes.node,
+    PropTypes.arrayOf(
+      PropTypes.oneOfType([
+        PropTypes.element,
+        PropTypes.arrayOf(PropTypes.element)
       ])
     )
   ]).isRequired,
-  zLevel: React.PropTypes.number.isRequired
+  zLevel: PropTypes.number.isRequired
 };
 
 Elevation.defaultProps = {
