@@ -1,6 +1,7 @@
 import React from 'react';
 import pick from 'lodash.pick';
 import styled from 'styled-components';
+import PropTypes from 'prop-types';
 
 import { propTypes as challengeProps } from '../../../../data/challenge';
 import { propTypes as disciplineActionProps } from '../../../../data/discipline_action';
@@ -40,11 +41,11 @@ const MemberProfile = props => {
 };
 
 MemberProfile.propTypes = {
-  canChallenge: React.PropTypes.bool.isRequired,
-  currentChallenge: React.PropTypes.shape(challengeProps),
-  currentDisciplineAction: React.PropTypes.shape(disciplineActionProps),
-  nextPerformance: React.PropTypes.shape(performanceProps.performance),
-  user: React.PropTypes.shape(userProps).isRequired
+  canChallenge: PropTypes.bool.isRequired,
+  currentChallenge: PropTypes.shape(challengeProps),
+  currentDisciplineAction: PropTypes.shape(disciplineActionProps),
+  nextPerformance: PropTypes.shape(performanceProps.performance),
+  user: PropTypes.shape(userProps).isRequired
 };
 
 export default MemberProfile;
