@@ -1,4 +1,4 @@
-import { PropTypes } from 'react';
+import PropTypes from 'prop-types';
 
 import helpers from './helpers';
 import { propTypes as spotProps } from '../spot';
